requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
tortoise-orm>=0.21,<1.0
aiomysql>=0.2,<1.0
pydantic>=2.6,<3.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
openpyxl>=3.1,<4.0
